Refactor pricing serializers and views; move to uom module and remove deprecated files
