+++ ecommerce/views/__init__.py
@@ -2,12 +2,12 @@ from .cart import CartItemView
 from .cart import CartListCreateView
 from .categories import CategoryDetailView
 from .categories import CategoryListView
-from .pricing import UnitOfMeasureDetailView
-from .pricing import UnitOfMeasureListView
 from .products import ProductDetailView
 from .products import ProductListCreateView
 from .subcategories import SubCategoryDetailView
 from .subcategories import SubCategoryListView
+from .uom import UnitOfMeasureDetailView
+from .uom import UnitOfMeasureListView
 
 __all__ = [
     "CartItemView",
